Make Orchestrated::Proxy a better delegate
Prior to this commit, Orchestrated::Proxy was, by default, a subclass of
Object, which gets polluted by ActiveSupport (and other gems in the
Rails framework), resulting in Orchestrated::Proxy#method_missing not
getting called (and thus the Orchestration isn't created) if there is a
name collision between ActiveSupport's Object monkey patching and an
orchestrated method (specifically #load as show in the README example).
This commit alters the base class of Orchestrated::Proxy to BasicObject
as this class is far less polluted by gems and frameworks. Additionally,
Kernel is included for utility methods like `block_given?`.